backend/requirements/requirements-development.txt
-r base.txt

# Debugging
django-debug-toolbar==4.2.0

click>=8.0,<9.0
flake8-bugbear==21.3.2
flake8-django==1.1.1
flake8==3.9.0
pep8-naming==0.11.1


# Linting
black==23.7.0
autoflake==1.4
bandit==1.7.0
ruff==0.0.252

# Type-checking
django-stubs==1.8.0
django_stubs_ext==0.1.0
mypy==1.0.1
types-Markdown==3.3.0
types-beautifulsoup4==4.10.4
types-redis==4.5.1.3
types-requests==2.25.11
celery-stubs==0.1.3

# Testing
coverage==5.3
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-xdist==3.2.1
pytest==7.2.2
